refactor: 조교 관리 전역 타입·상수·유틸 경계 정리
