+++ src/app/(dashboard)/educators/assistants/_components/AssistantsStatsGrid.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 
-import type { AssistantsStatItem } from "@/app/(dashboard)/educators/assistants/_types/assistants.page.types";
+import type { AssistantsStatItem } from "@/types/assistants.view";
 import { Card, CardContent } from "@/components/ui/card";
 
 type AssistantsStatsGridProps = {
