+++ src/app/(dashboard)/educators/assistants/_components/AssistantsTable.tsx
@@ -1,7 +1,7 @@
 import {
   type Assistant,
   type AssistantsPagination,
-} from "@/app/(dashboard)/educators/assistants/_types/assistants";
+} from "@/types/assistants.view";
 import StatusLabel from "@/components/common/label/StatusLabel";
 import { Pagination } from "@/components/common/pagination/Pagination";
 import { Card, CardContent } from "@/components/ui/card";
