+++ src/app/(dashboard)/educators/assistants/_modals/ResourceLibraryModal.tsx
@@ -3,7 +3,7 @@ import { Search } from "lucide-react";
 import {
   type ResourceLibraryItem,
   type ResourceLibraryCategory,
-} from "@/app/(dashboard)/educators/assistants/_types/assistants";
+} from "@/types/assistants.view";
 import { Button } from "@/components/ui/button";
 import { Checkbox } from "@/components/ui/checkbox";
 import {
