+++ src/constants/assistants.constants.ts
@@ -1,0 +1,182 @@
+import type {
+  ActiveStatusFilter,
+  Assistant,
+  AssistantDetailDraft,
+  AssistantStatus,
+  ContractRecord,
+  ContractStatus,
+  ResourceLibraryItem,
+} from "@/types/assistants.view";
+
+export const DEFAULT_ACTIVE_STATUS_FILTER: ActiveStatusFilter = "근무중";
+
+export const ACTIVE_STATUS_FILTER_OPTIONS: readonly ActiveStatusFilter[] = [
+  "근무중",
+  "근무전",
+  "퇴사",
+  "전체",
+];
+
+export const editableStatusOptions = ["근무전", "근무중"] as const;
+
+export const assistants: Assistant[] = [
+  {
+    id: "1",
+    name: "김민수",
+    email: "minsu.kim@academy.com",
+    subject: "수학",
+    phone: "010-1234-5678",
+    className: "중2 수학 심화반",
+    task: "주간 테스트 채점",
+    memo: "오답 정리 리포트 제출 주기 확인 필요",
+    status: "근무중",
+    badge: "bg-emerald-400/20 text-emerald-200",
+  },
+  {
+    id: "2",
+    name: "이진은",
+    email: "jineun.lee@academy.com",
+    subject: "영어",
+    phone: "010-9876-5432",
+    className: "고1 영어 내신대비",
+    task: "단어 시험 감독",
+    memo: "학부모 문의 대응 기록 있음",
+    status: "근무중",
+    badge: "bg-emerald-400/20 text-emerald-200",
+  },
+  {
+    id: "3",
+    name: "박성호",
+    email: "sungho.park@academy.com",
+    subject: "과학",
+    phone: "010-5555-4444",
+    className: "초6 창의과학 실험",
+    task: "업무 배정 예정",
+    memo: "다음 학기 재계약 여부 확인 필요",
+    status: "퇴사",
+    badge: "bg-slate-500/20 text-slate-300",
+  },
+  {
+    id: "4",
+    name: "최유리",
+    email: "yuri.choi@academy.com",
+    subject: "국어",
+    phone: "010-1111-2222",
+    className: "중3 국어 문법특강",
+    task: "학생 상담",
+    memo: "토요일 보충수업 지원 가능",
+    status: "근무전",
+    badge: "bg-amber-300/20 text-amber-100",
+  },
+  {
+    id: "5",
+    name: "정우성",
+    email: "woosung.jung@academy.com",
+    subject: "과학",
+    phone: "010-3333-7777",
+    className: "고2 물리 개념완성",
+    task: "업무 배정 예정",
+    memo: "정기 미팅 후 역할 재배정 예정",
+    status: "퇴사",
+    badge: "bg-slate-500/20 text-slate-300",
+  },
+];
+
+export const createAssistantDetailDraft = (
+  assistant?: Assistant
+): AssistantDetailDraft => ({
+  status: assistant?.status ?? "근무중",
+  memo: assistant?.memo ?? "",
+});
+
+export const statusColorMap: Record<
+  AssistantStatus,
+  "green" | "yellow" | "gray"
+> = {
+  근무전: "yellow",
+  근무중: "green",
+  퇴사: "gray",
+};
+
+export const contractRecords: ContractRecord[] = [
+  {
+    id: "contract-1",
+    assistantId: "1",
+    updatedAt: "2024-06-18",
+    fileName: "2024-1학기_김민수_계약서.pdf",
+    status: "서명 완료",
+  },
+  {
+    id: "contract-2",
+    assistantId: "2",
+    updatedAt: "2024-06-17",
+    fileName: "2024-1학기_이진은_계약서.pdf",
+    status: "서명 대기",
+  },
+  {
+    id: "contract-3",
+    assistantId: "3",
+    updatedAt: "2024-06-16",
+    fileName: "2024-1학기_박성호_계약서.pdf",
+    status: "재전송 필요",
+  },
+];
+
+export const contractStatusClassMap: Record<ContractStatus, string> = {
+  "서명 완료": "bg-emerald-100 text-emerald-700",
+  "서명 대기": "bg-amber-100 text-amber-700",
+  "재전송 필요": "bg-rose-100 text-rose-700",
+};
+
+export const contractTemplateOptions = [
+  "표준 근로계약서 (2024)",
+  "단기 근로계약서 (파트타임)",
+  "조교 계약 연장 양식",
+] as const;
+
+export const resourceLibraryItems: ResourceLibraryItem[] = [
+  {
+    id: "resource-1",
+    title: "중2 수학 1학기 기말고사 채점 가이드.pdf",
+    category: "평가자료",
+    updatedAt: "2026-02-07",
+    sizeLabel: "1.3MB",
+  },
+  {
+    id: "resource-2",
+    title: "중2 수학 오답노트 템플릿.hwp",
+    category: "수업자료",
+    updatedAt: "2026-02-06",
+    sizeLabel: "842KB",
+  },
+  {
+    id: "resource-3",
+    title: "조교 업무 체크리스트_학기중.xlsx",
+    category: "운영문서",
+    updatedAt: "2026-02-04",
+    sizeLabel: "560KB",
+  },
+  {
+    id: "resource-4",
+    title: "고1 영어 단어 테스트 감독 매뉴얼.pdf",
+    category: "수업자료",
+    updatedAt: "2026-02-03",
+    sizeLabel: "2.1MB",
+  },
+  {
+    id: "resource-5",
+    title: "모의고사 감독 유의사항 안내문.docx",
+    category: "평가자료",
+    updatedAt: "2026-01-31",
+    sizeLabel: "390KB",
+  },
+];
+
+export const resourceCategoryOptions = [
+  "전체",
+  "수업자료",
+  "평가자료",
+  "운영문서",
+] as const;
+
+export const PAGE_LIMIT = 5;
