+++ src/utils/assistants.ts
@@ -1,0 +1,30 @@
+const decodeHtmlEntities = (value: string) =>
+  value
+    .replace(/&nbsp;/gi, " ")
+    .replace(/&amp;/gi, "&")
+    .replace(/&lt;/gi, "<")
+    .replace(/&gt;/gi, ">")
+    .replace(/&quot;/gi, '"')
+    .replace(/&#39;/gi, "'");
+
+export const htmlToPlainText = (value?: string | null) => {
+  if (!value) {
+    return "";
+  }
+
+  const withLineBreaks = value
+    .replace(/<br\s*\/?>/gi, "\n")
+    .replace(/<\/p>\s*<p[^>]*>/gi, "\n")
+    .replace(/<p[^>]*>/gi, "")
+    .replace(/<\/p>/gi, "\n");
+
+  const withoutTags = withLineBreaks.replace(/<[^>]+>/g, "");
+  const decoded = decodeHtmlEntities(withoutTags);
+
+  return decoded
+    .split("\n")
+    .map((line) => line.trim())
+    .join("\n")
+    .replace(/\n{3,}/g, "\n\n")
+    .trim();
+};
